refactor qasereporter into 2 subcommands, rename the binary, update build files and docs
