+++ Dockerfile
@@ -31,11 +31,11 @@ RUN go mod download && \
 RUN cd $WORKSPACE && \
     make && \
     mv ./dartboard /usr/local/bin/dartboard && \
-    mv ./qasereporter-k6/qasereporter-k6 /usr/local/bin/qasereporter-k6
+    mv ./qasereporter-k6/qase-k6-cli /usr/local/bin/qase-k6-cli
 
 FROM grafana/k6:${K6_VERSION}@${K6_IMAGE_DIGEST}
 COPY --from=builder /usr/local/bin/dartboard /bin/dartboard
-COPY --from=builder /usr/local/bin/qasereporter-k6 /bin/qasereporter-k6
+COPY --from=builder /usr/local/bin/qase-k6-cli /bin/qase-k6-cli
 
 # Run the following commands as root user so that we can easily install some needed tools
 USER root
