+++ internal/qase/status.go
@@ -1,10 +1,11 @@
 package qase
 
 const (
-	StatusPassed  = "passed"
-	StatusFailed  = "failed"
-	StatusBlocked = "blocked"
-	StatusSkipped = "skipped"
-	StatusInvalid = "invalid"
-	StatusError   = "error"
+	StatusPassed             = "passed"
+	StatusFailed             = "failed"
+	StatusBlocked            = "blocked"
+	StatusSkipped            = "skipped"
+	StatusInvalid            = "invalid"
+	StatusError              = "error"
+	StatusExceededThresholds = "exceeded-thresholds"
 )
